+++ src/utils/torch_utils.py
@@ -24,7 +24,7 @@ class _ScaleGradient(torch.autograd.Function):
         if isinstance(scale, dict):
             scale = scale["value"]
 
-        return grad_output * scale, None
+        return (grad_output * scale).to(grad_output.dtype), None
 
 def scale_gradient(
     x: torch.Tensor,
